+++ src/containers/Login.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+
+import LoginForm from '../components/LoginForm'
+
+import './Login.css';
+
+class Login extends Component {
+  render() {
+    return (
+      <div>
+        <LoginForm />
+      </div>
+    );
+  }
+}
+
+export default Login;
